feat: add buffer size limits to protocol
Currently the largest message we might send is LIST_MESSAGES,
which can span up to 16MB (1+3+2^24). With the maximum size of a
message being 1100 bytes (33+33+1026), this would require a minimum
of 15,253 message objects to be sent. Realistically the server shouldn't
come close to this, and is more likely to be sending a few hundred
messages at a time. We will use a default buffer size of 1MiB which
should be suitable for all practical purposes.